:gem: Support for the continuation API. (#47)
New Butt#query method. It takes the params, like post, the return value to start with (defaults to an empty array, is sometimes an empty hash, or other things), and a block. The block passes the base return value, and everything in the query key in the API response. The return values modifications persist over the continuation loop.
Continuation can be enabled by setting use_continuation to true, either by directly setting the instance attribute, or by passing that to the options in initialize.
Modify the majority of the query methods to use the new continuation API. This removes many lines of redundant code :tada:

New Butt#query_ary helper method as described in PR #47.