:gem: Implement token caching (#75)
* :gem: :racehorse: Implement token caching
This is a significant performance improvement. Rather than retrieving a new token every time get_token is called, we have an internal cache of our tokens for each type. For the first badtoken error, we will try to retrieve a new token; subsequent badtokens will error (avoiding an infinite loop).
There is a feature degradation here as well, unfortunately. get_token used to be able to take an array of token types to get; with caching, this would be overly complicated (some tokens could be already cached, some might not, etc.), so instead the get_token feature now only takes a single token type. This should not be hugely significant, as I highly doubt many if any were making use of this feature.

* :package: Change get_token type assignment style from ternary